Query object updated
Query object now automatically generates the associative string representing variable types in the constructor for the purposes of binding variables. This now means you only need to pass the bare essentials to it.

Also you can pass single parameters to it without needing to wrap them in an array. Only need to use array() function when there is more than one parameter to be passed to the query.